requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.2
django-crispy-forms==2.0
django-recaptcha==3.0.0
django-tinymce==3.6.1
fontawesomefree==6.4.0
sqlparse==0.4.4
typing_extensions==4.7.0
